Worked on the IpoptApplication class (see ampl_ipopt) for an example of how to use the new IpoptApplication. Later we will switch the other interfaces to use the IpoptApplication class.
Added a FinalizeSolution method to NLP (and finalize_solution to TNLP)
which is called by the IpoptApplication following a successful solve.
We will need to change the return code stuff when we decide on a set of
return codes from the IpoptAlgorithm. Then, the algorithm can call
this FinalizeSolution method instead of the application (the IpoptAlgorithm
can also catch the exceptions, etc...)

Cheers,

Carl.


git-svn-id: https://projects.coin-or.org/svn/Ipopt/trunk@56 ef2a96d0-92f8-0310-a3c1-cdaa70078a94