Introduced initial structure for parameter handling in Ipopt.
The parameters are stored in the IpoptNLP. Code written for the old interface will still work the same way.
For now, the parameters are only read from a function of the user's TNLP implementation and saved
in a vector in the OrigIpoptNLP.
The next step will be pass these parameters back to the user's function evaluations.